requirements/base.txt
# Base
# not supported by django-reversion-compare==0.7.5
# django-reversion-compare==0.7.5
# diff-match-patch==20121119

Django==1.11.9
# celery==3.1.25
celery==4.1.0
geoip==1.3.2
django-ckeditor==5.2.2
django-db-mailer==2.3.19
django-docker-helpers>=0.1.8
django-reversion==2.0.8
django-dbtemplates==2.0

django-modeltranslation==0.12.1

pyaml
hiredis
django-redis-cache
dpath
gunicorn
uwsgi
django-uwsgi
raven
psycopg2
httpagentparser
django-ipware
whitenoise
